requirements.txt
python-telegram-bot==20.7
python-telegram-bot[callback-data]
python-dotenv
psycopg2-binary
aiofiles

